% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rccthemes-package.R
\docType{package}
\name{rccthemes-package}
\alias{rccthemes}
\alias{rccthemes-package}
\title{rccthemes: RCC Themes, Scales and Geoms for 'ggplot2'}
\description{
The longterm goal is to provide 'ggplot2'
    themes, scales and geoms that adhere to the graphic rules at Regional
    Cancer Centers (RCC) in Sweden.  Currently, RCC scales are
    implemented.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cancercentrum.bitbucket.io/rccthemes/}
  \item \url{http://bitbucket.org/cancercentrum/rccthemes}
  \item Report bugs at \url{https://bitbucket.org/cancercentrum/rccthemes/issues}
}

}
\author{
\strong{Maintainer}: John Lövrot \email{john.lovrot@sll.se} (\href{https://orcid.org/0000-0002-9339-8059}{ORCID})

Other contributors:
\itemize{
  \item Hadley Wickham (code from the ggplot2 and scales packages) [contributor]
  \item RCC [funder]
}

}
\keyword{internal}
