% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-rcc.R, R/zxx.R
\name{scale_colour_rcc}
\alias{scale_colour_rcc}
\alias{scale_fill_rcc}
\alias{scale_color_rcc}
\title{RCC colours for discrete data}
\usage{
scale_colour_rcc(
  ...,
  aesthetics = "colour",
  direction = 1,
  na.value = "grey50"
)

scale_fill_rcc(..., aesthetics = "fill", direction = 1, na.value = "grey50")
}
\arguments{
\item{...}{Other arguments passed on to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}} to
control name, limits, breaks, labels and so forth.}

\item{aesthetics}{Character string or vector of character strings listing the
name(s) of the aesthetic(s) that this scale works with. This can be useful, for
example, to apply colour settings to the \code{colour} and \code{fill} aesthetics at the
same time, via \code{aesthetics = c("colour", "fill")}.}

\item{direction}{Sets the order of colours in the scale. If 1, the default,
colours are as the in RCC graphic profile. If -1, the order of
colours is reversed.}

\item{na.value}{Colour to use for missing values}
}
\description{
Colour scale for categorical variables based on RCC graphic profile.
}
\examples{
library(ggplot2)
library(rccthemes)
theme_set(theme_minimal())

# Example of using the RCC scale for the esthetic colour
ggplot(mpg, aes(x = displ, y = hwy, colour = class)) +
  geom_point() +
  scale_colour_rcc()

# Example of using the RCC scale for the esthetic fill
g <- ggplot(mpg, aes(x = class, fill = drv)) +
  geom_bar()

g + scale_fill_rcc()
g + scale_fill_rcc(direction = -1)
}
\references{
\url{https://bitbucket.org/cancercentrum/rccshiny/src/11416e96246ed54c1b60497c20510ae842eb260a/R/fIndPlot.R?at=master&fileviewer=file-view-default#fIndPlot.R-291}
}
\concept{colour scales}
