% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-rcc.R
\name{rcc_pal}
\alias{rcc_pal}
\title{RCC palette (discrete).}
\usage{
rcc_pal(direction = 1)
}
\arguments{
\item{direction}{Sets the order of colours in the scale. If 1, the default,
colours are as the in RCC graphic profile. If -1, the order of
colours is reversed.}
}
\description{
RCC palette (discrete).
}
\examples{
library(scales)
library(rccthemes)
show_col(rcc_pal()(12))
}
\references{
\url{https://bitbucket.org/cancercentrum/rccshiny/src/1.10.0/R/rcc2PlotLine.R#lines-120}
}
